마우스 클릭 이펙트 적용

+++ Assets/03.PrivateFolder/KangMS/KMS_Scripts/01_Note/DBScoreNote.cs
@@ -22,7 +22,7 @@ public class DBScoreNote : Note, IPoolingObj
 
     public E_Pool MyPoolType => E_Pool.DBSCORE_NOTE;
 
-    public override void Initialize(Vector3 endPoint, float speed, int damage = 0, int stageNumber = 1, E_NoteType noteType = E_NoteType.None, E_SpawnerPosY notePosition = E_SpawnerPosY.BOTTOM)
+    public override void Initialize(Vector3 endPoint, float speed, float damage = 0, int stageNumber = 1, E_NoteType noteType = E_NoteType.None, E_SpawnerPosY notePosition = E_SpawnerPosY.BOTTOM)
     {
         base.Initialize(endPoint, speed, damage, stageNumber, noteType, notePosition);
